switch to DoubleNumerics

+++ PTSharpCore/Function.cs
@@ -8,7 +8,7 @@ namespace PTSharpCore
 {
     interface Func : IShape
     {
-        float func(float x, float y);
+        double func(double x, double y);
     }
     
     class Function : Func
@@ -45,9 +45,9 @@ namespace PTSharpCore
 
         Hit IShape.Intersect(Ray ray)
         {
-            float step = 1.0F / 32F;
+            double step = 1.0F / 32F;
             bool sign = Contains(ray.Position(step));
-            for (float t = step; t < 12; t += step)
+            for (double t = step; t < 12; t += step)
             {
                 V v = ray.Position(t);
                 if (Contains(v) != sign && Box.Contains(v))
@@ -60,12 +60,12 @@ namespace PTSharpCore
 
         V IShape.UV(V p)
         {
-            float x1 = Box.Min.v.X;
-            float x2 = Box.Max.v.X;
-            float y1 = Box.Min.v.Y;
-            float y2 = Box.Max.v.Y;
-            float u = p.v.X - x1 / x2 - x1;
-            float v = p.v.Y - y1 / y2 - y1;
+            double x1 = Box.Min.v.X;
+            double x2 = Box.Max.v.X;
+            double y1 = Box.Min.v.Y;
+            double y2 = Box.Max.v.Y;
+            double u = p.v.X - x1 / x2 - x1;
+            double v = p.v.Y - y1 / y2 - y1;
             return new V(u, v, 0);
         }
 
@@ -76,15 +76,15 @@ namespace PTSharpCore
 
         V IShape.NormalAt(V p)
         {
-            float eps = 1e-3F;
-            float x = func(p.v.X - eps, p.v.Y) - func(p.v.X + eps, p.v.Y);
-            float y = func(p.v.X, p.v.Y - eps) - func(p.v.X, p.v.Y + eps);
-            float z = 2 * eps;
+            double eps = 1e-3F;
+            double x = func(p.v.X - eps, p.v.Y) - func(p.v.X + eps, p.v.Y);
+            double y = func(p.v.X, p.v.Y - eps) - func(p.v.X, p.v.Y + eps);
+            double z = 2 * eps;
             V v = new V(x, y, z);
             return v.Normalize();
         }
 
-        public float func(float x, float y)
+        public double func(double x, double y)
         {
             return Funct.func(x, y);
         }
