+++ PTSharpCore/ITexture.cs
@@ -2,11 +2,11 @@
 {
     interface ITexture
     {
-        Colour Sample(float u, float v);
-        V NormalSample(float u, float v);
-        V BumpSample(float u, float v);
-        ITexture Pow(float a);
-        ITexture MulScalar(float a);
+        Colour Sample(double u, double v);
+        V NormalSample(double u, double v);
+        V BumpSample(double u, double v);
+        ITexture Pow(double a);
+        ITexture MulScalar(double a);
     }
 }
