+++ PTSharpCore/Material.cs
@@ -9,15 +9,15 @@ namespace PTSharpCore
         public ITexture NormalTexture;
         public ITexture BumpTexture;
         public ITexture GlossTexture;
-        public float BumpMultiplier;
-        public float Emittance;
-        public float Index;
-        public float Gloss;
-        public float Tint;
-        public float Reflectivity;
+        public double BumpMultiplier;
+        public double Emittance;
+        public double Index;
+        public double Gloss;
+        public double Tint;
+        public double Reflectivity;
         public bool Transparent;
 
-        public Material(Colour color, ITexture texture, ITexture normaltexture, ITexture bumptexture, ITexture glosstexture, float b, float e, float i, float g, float tint, float r, Boolean t)
+        public Material(Colour color, ITexture texture, ITexture normaltexture, ITexture bumptexture, ITexture glosstexture, double b, double e, double i, double g, double tint, double r, Boolean t)
         {
             Color = color;
             Texture = texture;
@@ -38,32 +38,32 @@ namespace PTSharpCore
             return new Material(color, null, null, null, null, 1, 0, 1, 0, 0, -1, false);
         }
 
-        public static Material SpecularMaterial(Colour color, float index)
+        public static Material SpecularMaterial(Colour color, double index)
         {
             return new Material(color, null, null, null, null, 1, 0, index, 0, 0, -1, false);
         }
 
-        public static Material GlossyMaterial(Colour color, float index, float gloss)
+        public static Material GlossyMaterial(Colour color, double index, double gloss)
         {
             return new Material(color, null, null, null, null, 1, 0, index, gloss, 0, -1, false);
         }
 
-        public static Material ClearMaterial(float index, float gloss)
+        public static Material ClearMaterial(double index, double gloss)
         {
             return new Material(new Colour(0, 0, 0), null, null, null, null, 1, 0, index, gloss, 0, -1, true);
         }
 
-        public static Material TransparentMaterial(Colour color, float index, float gloss, float tint)
+        public static Material TransparentMaterial(Colour color, double index, double gloss, double tint)
         {
             return new Material(color, null, null, null, null, 1, 0, index, gloss, tint, -1, true);
         }
 
-        public static Material MetallicMaterial(Colour color, float gloss, float tint)
+        public static Material MetallicMaterial(Colour color, double gloss, double tint)
         {
             return new Material(color, null, null, null, null, 1, 0, 1, gloss, tint, 1, false);
         }
 
-        public static Material LightMaterial(Colour color, float emittance)
+        public static Material LightMaterial(Colour color, double emittance)
         {
             return new Material(color, null, null, null, null, 1, emittance, 1, 0, 0, -1, false);
         }
